feat(common): Change `int` to `long` when creating `newInt`:

+++ src/main/java/com/abiddarris/common/renpy/internal/Python.java
@@ -49,7 +49,7 @@ public class Python {
         return object;
     }
     
-    public static PythonObject newInt(int value) {
+    public static PythonObject newInt(long value) {
         PythonObject object = new PythonInt(value);
         object.setAttribute("__class__", int0);
